Handle zone conflict in purge mode
If managing a purge vent, monitoring 2 zones, it is possible that the two zones' thermostatOperatingState will conflict. if one is fan-only and the other is heat/cool, we want to react to the heat/cool action. But if they are conflicted (one is heating, other is cooling) we really don't know what to do, so we go into idle mode (for now).

Note that this conflict is possible with the use of a zone controller such as the AprilAire 6202/6203, which will arbitrate the actual demand such that one is serviced at a time. Unfortunately, we have no way of knowing which one is serviced first (we will know which is serviced second, because the first will go idle once it has been serviced).

Note to self: consider following the temperature change at both thermostats to figure out which is being serviced...is the $cooling one getting cooler, or the $heating one getting warmer?